+++ src/lesson9/homework/ManageCars.java
@@ -1,0 +1,11 @@
+package lesson9.homework;
+
+public class ManageCars {
+    public static void main(String[] args) {
+        //Создать любой класс по желанию и реализовать в нем минимум 2 метода.
+        Cars bmw = new Cars();
+        bmw.printCarsInfo(" тыщ денег");
+        Cars lincoln = new Cars("Линкольн", 2.0, 15000);
+        lincoln.printCarsInfo("");
+    }
+}
